Adds to cart (variable)
Adds and removes items from cart variable feature,
Now also changed the recycler view to respond to user onclicks, when the user clicks the recycler item it would show the item page asssosciated with the recyclerview that was clicked.
(Still working on retrieving the data from DB and showing it in the page).