Main Menu Products
Now if you go to main menu, the products would show up, and would be added from the data bases if needed, or otherwise would always generate a default list of product.

When clicks on product, it will show him the details page about it.

When he adds or subtract he can go to cart and view how many there are of it.

Detailed information about product also shown in the main page.